Comeca a transformar o env em lista encadeada
