+++ includes/minishell.h
@@ -42,6 +42,7 @@ typedef struct s_env
 {
 	char			*key;
 	char			*value;
+	char			*full;
 	struct s_env	*next;
 }					t_env;
 
@@ -72,4 +73,12 @@ typedef struct s_env
 # define MINISHELL	P R O M PP T BOLD_GREY"$> "RESET
 //---------------------------------END MACROS---------------------------------//
 
+/*__________________________________FUNCTIONS____________________________________*/
+
+char		*ft_strdup_equal_value(const char *src);
+char		*ft_strdup_equal_key(const char *src);
+char		**get_key(char **envp);
+char		**get_value(char **envp);
+t_env_arr	*init_env(char **envp);
+//---------------------------------END FUNCTIONS---------------------------------//
 #endif
